backend/requirements.txt
python-dotenv
sqlalchemy
passlib[bcrypt]
pyjwt